backend/requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-CORS==4.0.0

# Database
pymongo==4.5.0

# SMS Service
twilio==8.9.1

# Machine Learning
pandas==2.1.1
numpy==1.25.2
scikit-learn==1.3.0
joblib==1.3.2

# Scheduling
schedule==1.2.0

# Environment variables
python-dotenv==1.0.0

# Data processing (for ML models)
scipy==1.11.3

# Optional: For enhanced data handling
openpyxl==3.1.2
xlsxwriter==3.1.9

# Development dependencies (optional)
pytest==7.4.2
pytest-flask==1.2.0
black==23.9.1
flake8==6.1.0

# Production WSGI server (optional)
gunicorn==21.2.0